chore : cookie 관련 유틸함수들 하나의 파일로 이동
