--- src/utils/auth/checkCookie.ts
@@ -1,9 +1,0 @@
-'use server';
-
-import { cookies } from 'next/headers';
-
-export const checkCookie = (key: string) => {
-  return cookies().has(key);
-};
-
-export default checkCookie;
