--- src/utils/auth/getCookie.ts
@@ -1,9 +1,0 @@
-'use server';
-
-import { cookies } from 'next/headers';
-
-export const getCookie = (key: string) => {
-  return cookies().get(key)?.value;
-};
-
-export default getCookie;
