--- src/utils/auth/setCookie.ts
@@ -1,9 +1,0 @@
-"use server";
-
-import { cookies } from "next/headers";
-
-export const setCookie = (key: string, value: string) => {
-  return cookies().set(key, value);
-};
-
-export default setCookie;
